Created api using prisma

+++ src/app/api/v1/users/[id]/route.ts
@@ -26,6 +26,16 @@ export async function DELETE(
   const { id } = params;
 
   try {
+    const userId = await prisma.user.findFirst({
+      where: { id: Number(id) },
+    });
+
+    if (!userId) {
+      return NextResponse.json({
+        message: "User does not exists",
+        success: false,
+      });
+    }
     await prisma.users.delete({
       where: { id: Number(id) },
     });
@@ -42,6 +52,17 @@ export async function PUT(
   const { id } = params;
   try {
     const { name } = await request.json();
+
+    const userId = await prisma.user.findFirst({
+      where: { id: Number(id) },
+    });
+
+    if (!userId) {
+      return NextResponse.json({
+        message: "User does not exists",
+        success: false,
+      });
+    }
 
     await prisma.user.update({
       where: { id: Number(id) },
